refactor(ui): place transcript types with their owners
